editoast: move Sign to editoast_schemas
