+++ editoast/editoast_schemas/src/infra.rs
@@ -1,6 +1,7 @@
 mod direction;
 mod directional_track_range;
 mod side;
+mod sign;
 mod track_location;
 mod track_offset;
 mod track_range;
@@ -9,6 +10,7 @@ mod waypoint;
 pub use direction::Direction;
 pub use directional_track_range::DirectionalTrackRange;
 pub use side::Side;
+pub use sign::Sign;
 pub use track_location::TrackLocation;
 pub use track_offset::TrackOffset;
 pub use track_range::TrackRange;
