+++ editoast/src/schema/neutral_section.rs
@@ -4,9 +4,9 @@ use serde::Serialize;
 
 use super::OSRDIdentified;
 use super::ObjectType;
-use super::Sign;
 use editoast_common::Identifier;
 use editoast_schemas::infra::DirectionalTrackRange;
+use editoast_schemas::infra::Sign;
 use editoast_schemas::primitives::OSRDTyped;
 
 /// Neutral sections are portions of track where trains aren't allowed to pull power from electrifications. They have to rely on inertia to cross such sections.
