+++ editoast/src/schema/speed_section.rs
@@ -7,11 +7,11 @@ use serde::Serialize;
 use super::ApplicableDirectionsTrackRange;
 use super::OSRDIdentified;
 use super::ObjectType;
-use super::Sign;
 use crate::infra_cache::Cache;
 use crate::infra_cache::ObjectCache;
 use editoast_common::Identifier;
 use editoast_common::NonBlankString;
+use editoast_schemas::infra::Sign;
 use editoast_schemas::primitives::OSRDTyped;
 
 #[derive(Debug, Derivative, Clone, Serialize, PartialEq, Copy)]
